% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{add_no_answer_level}
\alias{add_no_answer_level}
\title{Add a "no answer" level}
\usage{
add_no_answer_level(data, levels)
}
\arguments{
\item{data}{survey replies with a value of 0, which equals "no answer"}

\item{levels}{Adding a new "No answer level" to a vector containing all levels}
}
\description{
If there are individuals with no answer, add the "no answer" level
}
\examples{
add_no_answer_level(data, Q1_levels)

}
