% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{format_question_likert}
\alias{format_question_likert}
\title{Create data frame for plotting likert questions}
\usage{
format_question_likert(sod_data, pigweb_data, levels)
}
\arguments{
\item{sod_data}{a vector of answers from the sod_data data frame that contains the answers to the specific question}

\item{pigweb_data}{a vector of answers from the pigw_data data frame that contains the answers to the specific question}

\item{levels}{A vector of the levels of response categories in order from low to high}
}
\description{
Create a data frame for plotting that combines SoD and PigWeb data
}
\examples{
format_question_likert(sod21, pigw21, Q1_levels)

}
