% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{format_question_likert_pigweb}
\alias{format_question_likert_pigweb}
\title{Create a data frame for plotting that only include PigWeb data}
\usage{
format_question_likert_pigweb(pigweb_data, levels)
}
\arguments{
\item{pigweb_data}{vector of answers from the data frame that contains the answers from the PigWeb 2021 survey}

\item{levels}{vector of the levels of response categories in order from low to high}
}
\description{
Create a data frame for plotting that only include PigWeb data
}
\examples{
format_question_likert_pigweb(pigw21Q14$Q14, Q14_levels)

}
