% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{format_question_multiple_choice_pigweb}
\alias{format_question_multiple_choice_pigweb}
\title{doCreate a data frame for plotting for multiple-choice questions only included}
\usage{
format_question_multiple_choice_pigweb(survey_data, levels)
}
\arguments{
\item{survey_data}{vector of answers from the data frame that contains the answers from the PigWeb 2021 survey}

\item{levels}{vector of the levels of response categories}
}
\description{
Create a data frame for plotting for multiple-choice questions only included
}
\examples{
format_question_multiple_choice_pigweb(pigw21, Q15_levels)

}
