% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{format_question_numeric}
\alias{format_question_numeric}
\title{Create a data frame for plotting numeric questions}
\usage{
format_question_numeric(sod_data, pigweb_data)
}
\arguments{
\item{sod_data}{vector of answers from the data frame that contains the answers from the SOD 2021 survey}

\item{pigweb_data}{vector of answers from the data frame that contains the answers from the PigWeb 2021 survey}
}
\description{
Create a data frame for plotting numeric questions
}
\examples{
format_question_numeric(sod21, pigw21)

}
