% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{format_question_single_choice}
\alias{format_question_single_choice}
\title{Create a data frame for plotting for single-choice questions}
\usage{
format_question_single_choice(sod_data, pigweb_data, levels)
}
\arguments{
\item{sod_data}{vector of answers from the data frame that contains the answers from the SOD 2021 survey}

\item{pigweb_data}{vector of answers from the data frame that contains the answers from the PigWeb 2021 survey}

\item{levels}{vector of the levels of response categories}
}
\description{
Create a data frame for plotting for single-choice questions
}
\examples{
format_question_single_choice(sod21, pigw21, Q3_levels)

}
