% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{format_question_single_choice_pigweb}
\alias{format_question_single_choice_pigweb}
\title{Create a data frame for plotting for single-choice questions}
\usage{
format_question_single_choice_pigweb(pigweb_data, levels)
}
\arguments{
\item{pigweb_data}{vector of answers from the data frame that contains the answers from the PigWeb 2021 survey}

\item{levels}{vector of the levels of response categories}
}
\description{
Create a data frame for plotting for single-choice questions
}
\examples{
format_question_single_choice_pigweb(pigw21, Q9_levels)

}
