% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{plot_likert}
\alias{plot_likert}
\title{Plot a cumulative step graph of the answers for a Likert-scale question}
\usage{
plot_likert(data, levels, question, xlab)
}
\arguments{
\item{data}{data frame produced by function format_question_likert()}

\item{levels}{vector of the levels of response categories in order from low to high}

\item{question}{character string of the question, will be the title of the plot}

\item{xlab}{character string for the x-axis label}
}
\description{
Create a cumulative step graph of the answers for a Likert-scale question
}
\examples{
plot_likert(Q1_data, Q1_levels, Q1,"Time since most recent publication")

}
