% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{plot_multiple_choice}
\alias{plot_multiple_choice}
\title{Plot a bar chart of percentages for a multiple response question}
\usage{
plot_multiple_choice(data, levels, question, xlab)
}
\arguments{
\item{data}{data frame produced by function format_question_multiple_choice()}

\item{levels}{vector of the levels of response categories}

\item{question}{character string of the question, will be the title of the plot}

\item{xlab}{character string for the x-axis label}
}
\description{
Create a bar chart of percentages for a multiple response question
}
\examples{
plot_multiple_choice(Q4_data, Q4_levels, Q4, "Response")

}
