% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{plot_numeric}
\alias{plot_numeric}
\title{Plot a cumulative step graph of the answers for a numeric question}
\usage{
plot_numeric(data, question, xlab)
}
\arguments{
\item{data}{data frame produced by function format_question_numeric()}

\item{question}{character string of the question, will be the title of the plot}

\item{xlab}{character string for the x-axis label}
}
\description{
Create a cumulative step graph of the answers for a numeric question
}
\examples{
plot_numeric(Q2_data, Q2, "Year of first publication by respondent")

}
