% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{plot_single_choice}
\alias{plot_single_choice}
\title{Plot a bar chart of percentages for a single response question}
\usage{
plot_single_choice(data, levels, question, xlab)
}
\arguments{
\item{data}{data frame produced by function format_question_single_choice()}

\item{levels}{vector of the levels of response categories}

\item{question}{character string of the question, will be the title of the plot}

\item{xlab}{character string for the x-axis label}
}
\description{
Create a bar chart of percentages for a single response question
}
\examples{
plot_single_choice(Q3_data, Q3_levels, Q3, "Area of interest")

}
