% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_package.R
\name{reverseCode}
\alias{reverseCode}
\title{Reverses the order of Likert scale question responses.}
\usage{
reverseCode(x, min = 1, max = 5)
}
\arguments{
\item{x}{is a scalar from or a vector of Likert scores to reverse code}

\item{min}{is the minimum value of the Likert scale}

\item{max}{is the maxium value of the Likert scale}
}
\description{
Reverses the order of Likert scale question responses.
}
\examples{
reverseCode(2, min = 1, max = 5) # reverse code "2" to a "4" on a Likert scale of 1-5
reverseCode(2:7, min = 2, max = 7) # reverse code a vector of scores on a Likert scale of 2-7
reverseCode(1, min = 0, max = 1) # reverse code binary response

}
